Rename Accordion/index.jsx
