+++ src/index.js
@@ -1,6 +1,6 @@
 import './main.scss';
 
-export { default as Accordion } from './components/Accordion';
+export { default as Accordion } from './components/Accordion/Accordion';
 export { default as BlockContent } from './components/BlockContent';
 export { default as Breadcrumb } from './components/Breadcrumb';
 export { default as Button } from './components/Button';
